pyproject.toml
[project]
name = "LDDC"
dynamic = ["version"]
dependencies = [
    "PySide6-Essentials>=6.8.0",
    "httpxhttpx[http2]",
    "mutagen",
    "diskcache",
    "charset-normalizer",
    "pyaes",
    "psutil",
    "pyobjc-framework-Cocoa; sys_platform == 'darwin'",
]
requires-python = ">= 3.10"
description = "A Practical Tool for Downloading verbatim Lyrics"
readme = "README.md"
maintainers = [
    {name = "沉默の金", email = "cmzj@cmzj.org"}
]
license = {file = "LICENSE"}
keywords = ["lyrics","music","download","acquisition","tool"]
classifiers = [
    "Intended Audience :: End Users/Desktop",

    "Environment :: MacOS X",
    "Environment :: Win32 (MS Windows)",
    "Environment :: X11 Applications :: Qt",

    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",

    "Natural Language :: Chinese (Simplified)",
    "Natural Language :: English",
    "Natural Language :: Japanese",

    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",

    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",

    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Utilities",
]

[project.urls]
homepage = "https://github.com/chenmozhijin/LDDC"
repository = "https://github.com/chenmozhijin/LDDC"
documentation = "https://github.com/chenmozhijin/LDDC/wiki"
tracker = "https://github.com/chenmozhijin/LDDC/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["LDDC/**"]
exclude = ["*.ui", "*.qrc", "LDDC/res/img", "LDDC/res/i18n"]

[tool.hatch.version]
path = "LDDC/utils/version.py"

[tool.ruff]
target-version = "py310"
line-length = 159

[tool.ruff.lint]
select = [
    "ALL",

    "W",
    "E",

    "CPY001",
]

ignore = [
    "ANN401",  # any-type
    "BLE001",  # blind-except
    "D100",  # undocumented-public-module
    "D101",  # undocumented-public-class
    "D102",  # undocumented-public-method
    "D103",  # undocumented-public-function
    "D104",  # undocumented-public-package
    "D105",  # undocumented-magic-method
    "D106",  # undocumented-public-nested-class
    "D107",  # undocumented-public-init
    "D400",  # ends-in-period
    "D415",  # ends-in-punctuation
    "ERA001",  # commented-out-code
    "PLR2004",  # magic-value-comparison
    "Q000",  # bad-quotes-inline-string
    "RUF001",  # ambiguous-unicode-character-string
    "N802",  # invalid-function-name
    "N999",  # invalid-name

    # "PTH",  # flake8-use-pathlib
    "FBT",  # flake8-boolean-trap
]
preview = true
explicit-preview-rules = true

[tool.ruff.lint.pylint]
max-branches = 25  # PLR0912
max-returns = 15  # PLR0911
max-statements = 75  # PLR0915
max-args = 10  # PLR0913

[tool.ruff.lint.mccabe]
max-complexity = 30  # C901

[tool.ruff.lint.extend-per-file-ignores]
"tests/**.py" = ["S101"]
"tools/**.py" = ["T201", "INP001"]